+++ meteor-app/imports/startup/client/startup.js
@@ -15,11 +15,11 @@ const setupStreamer = () => {
   if (Meteor.isClient) {
     sendMessage = function (message) {
       streamer.emit('message', message)
-      console.log('me: ' + message)
+      // console.log('me: ' + message)
     }
 
     streamer.on('message', function (message) {
-      console.log('user: ', message)
+      // console.log('user: ', message)
     })
     streamer.emit('message', 'message from client')
     streamer.emit('message', { a: 'message from client' })